Structured output, better cropping and zooming.
Output of the parse_file function is not a Document object which
contains the Paragraphs and Headings, but also a list of Page objects
which provide information about page sizes and how the pages were
cropped.

Provided support for controlling the "zooming" of the document.

Started work on having different crops for different pages, but not
part of the API yet.